requirements.txt
apturl==0.5.2
asgiref==3.6.0
attrs==22.2.0
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
click==8.0.3
colorama==0.4.4
command-not-found==0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.1
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.1
Django==4.1.7
django-elasticsearch-dsl==7.2.2
django-elasticsearch-dsl-drf==0.22.5
django-environ==0.9.0
django-filter==22.1
django-js-asset==2.0.0
django-mptt==0.14.0
django-nine==0.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
duplicity==0.8.21
elasticsearch==7.17.8
elasticsearch-dsl==7.4.0
exceptiongroup==1.0.4
Faker==15.3.4
fasteners==0.14.1
future==0.18.2
httplib2==0.20.2
idna==3.4
importlib-metadata==4.6.4
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
jeepney==0.7.1
Jinja2==3.1.2
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.1
model-bakery==1.9.0
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
packaging==22.0
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.3.0
pluggy==1.0.0
protobuf==3.12.4
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
PyGObject==3.42.1
PyJWT==2.6.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pytest==7.2.0
pytest-django==4.5.2
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
pytz==2022.7
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.4.3
systemd-python==234
tomli==2.0.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.13
usb-creator==0.3.7
wadllib==1.3.6
xdg==5
xkit==0.0.0
zipp==1.0.0
